JavaEngine: split Builder initialization logic
Rather than doing all the work in the constructor, and throwing a
bunch of exceptions, let's keep construction as simple as possible,
then do all that initialization in its own method. This will make
error handling more robust in that we will be able to attempt calls to
builder.cleanup() even if something went wrong during initialization.